Added a PHP Dev vagrant box that I built
This is a Cent 6.4 x86_64 minimal install with the addition of the latest stable Nginx, PHP-FPM 5.3 and 5.4. It also includes phpMyAdmin and Mailcatcher. Percona 5.5 is installed. The idea is to give a reasonable PHP DEV box with a quick boot time, the image is slightly larger than I wanted, but, boot time is around 20 seconds on my iMac. Chef and Puppet are also configured. It runs fine in 512MB of RAM. 

If you create a file 'php54' in your project directory (/vagrant) and restart 'service php-fpm restart' you can switch between PHP5.3 and PHP5.4 easily. Both versions use the same php.ini file and have the same './configure' line.